requirements.txt
pygame
pillow
tqdm
joblib
pyturbojpeg